[GNB Sim] Gunbreaker Sim (#429)
A Gunbreaker Sim. It's not perfect, but it performs everything admirably
and gets all of its important abilities in No Mercy.

For weird speeds where things don't align perfectly (e.g. 2.45), it opts
to clip (if the clip is caused by GCD speed) to clip to keep the GCDs
from drifting. This seems to be better than any alternative I could
find.


![image](https://github.com/user-attachments/assets/0a2a25b3-8682-497b-b4ea-7fa498f1411d)


![image](https://github.com/user-attachments/assets/fed55976-05b4-4a32-bf4e-5a4cd8b9abe6)