Correct Riddle auto timing in a hacky way
While talking with Hint, she expressed that it would be desirable to fix
RoW auto timing no matter how, as inaccurate auto timing would favor
higher SkS gearsets due to auto scaling.

This fix is mathematically valid, but obviously needs to be documented
in case changes to haste calculations are done in the future.

`2.56 * (100 - 40 - 20) / 100 == 1.024`