Add warning to meld solver for solving pentameld sets (#467)
Closes https://github.com/xiv-gear-planner/gear-planner/issues/374

This adds a warning to the meld solver when the number of melds it needs
to solve for is above 23 (the maximum number without pentameld items).
While it's an 'unscientific' number (24 melds would be easy to solve
for), it displays the warning when the user starts getting into
difficult territory.

It will dynamically check if 'Overwrite existing materia' is set and
display the warning if enabling this setting pushes it over the limit of
23.


![image](https://github.com/user-attachments/assets/4233c30f-2996-4a12-87c8-a59105153170)

![image](https://github.com/user-attachments/assets/ae44812d-2e53-4603-a113-87bc480d1c76)

![image](https://github.com/user-attachments/assets/f11367ad-24b1-4251-a054-3c530e9934e4)

![image](https://github.com/user-attachments/assets/a0fe2303-bd6d-4c88-bfd4-1bb027de6db5)