Always show level in sync info it item level is synced (#487)
This will always show level in the sync info if the item level is
synced. Mostly, this helps prevents silly mistakes like forgetting to
sync level and makes everything a lot more explicit, so I think this is
generally a positive improvement. I think future work would include
making 'strange' syncs' warn users, but I think this is a positive even
in that world.


![image](https://github.com/user-attachments/assets/5631e425-0f0c-40ae-8229-ed3720969d45)

(Works in view mode too)

`tsconfig-paths-webpack-plugin` seems to be required to build, and it
wasn't present in a package.json (npm install didn't install it) so I
added it.