Add more things for analytics (#521)
`openNormalBacklink` when you click the link to go back to the full
sheet when viewing an onlySetIndex link.

`openEmbedToFull` is the same but when you're looking at an embed.

`displayEmbedError` is whenever the function `displayEmbedError()` is
called.

`nonEmbeddedIframe` is fired when using a non-embedded UI in an iframe.

`iframe` is fired whenever the page is in an iframe.

#1 does not work for non-iframes because the page navigation immediately
cancels the request. This should be fixed, but is out of scope for this.
It works in iframes because the link is set to open in a new tab.