Make Meld Solver button first-class, rename materia fill/solve to materia fill/lock (#534)
There are a few problems with including the meld solver in the fill/lock
button in my eyes:
- Users get confused that the fill priority affects the meld solver
- This is one of the most awesome features of the site, and hiding it in
a sub-menu can cause people to not know about it

This PR makes the meld solver button first-class, and moves it out of
the fill/lock button. Making it more obviously disconnected from the
'fill' capabilities, as well as being more obvious to users.


![image](https://github.com/user-attachments/assets/1d2f1d9c-9bef-4976-bd42-f713b2868cf9)

![image](https://github.com/user-attachments/assets/88aa2838-3642-4981-98d6-83fd2daf424a)

![image](https://github.com/user-attachments/assets/175cd0fa-e036-4d0e-9ea1-ddd50759b4bd)