Implement the triple lunar rotation (#566)
Another rotation for long encounters. Now that rotations can be selected
by the end user I thought it would be valuable to have the 3rd monk
rotation displayed. Previously I didn't think it was worth including as
if it simmed the highest it would not be representative of "average
play".

Co-authored-by: JC <dev@jacobcasper.com>