Multi-job sheets (#586)
Closes #272 and #587

This allows you to have a role-based sheet, which includes jobs of all
roles. The sheet still has a "primary" job as selected on the new sheet
(or the "save as" screen), but you can change the job for an individual
set within the sheet.

- [x] Make Datamanager server support querying multiple jobs' worth of
items in one request.
- [x] Add multi-job logic to sheets and sets.
- [x] Make UI show appropriate items.
- [x] UI to pick job for an individual sheet.
- [x] Import/export logic.
- [x] Make sure it works correctly when you "Save As" the sheet for a
different class.
- [x] Make the "Save As" dialog only show jobs of the same role.