First attempt at improving predictive tiering (#620)
When lookin at predicted/extended tiering (i.e. tiering plus or minus a
certain amount of materia), this will now reflect the tiering as if you
were to actually equip the item, rather than just adding the raw amount
that the materia would provide. This provides a more realistic picture
when you have food with an uncapped stat.

Fixes #606