Sim table enhancements (#645)
Closes #561 by allowing tables to lazy evaluate rows that are not
visible on screen.

Implements this for existing cycle sim UIs. Also provides a method for
cycle sim UIs to directly specify their extra columns, rather than
needing to override the table method itself. The former was unperformant
and broke the lazy behavior because it was adding the columns after the
data was already set, which meant there were multiple re-evaluations of
the table.