Initial BLM sim implementation (#597)
Only level 100 for now but planning to add lower levels later. Also need
to clean up the code.
The way it weaves oGCDs is quite bruteforcey but it works since it's
just a dummy rotation.

The results themselves look pretty accurate* to the "7.X BLM Relative
Damage Calc" sheet for the few gearsets that I've tried which is
promising. *or at least in the same ballpark as them

---------

Co-authored-by: xpdota <smurfs@smurfs.verymad.net>