Updated BLM sim (#685)
Cleaned leftover comments/code from DRK sim

Updated to use the new gauge system instead of a custom gauge
Fixed Blizzard IV having a 3.5s cast time instead of the intended 2.0s
cast time

Some notes, I changed the element from being a number between -3 and +3
to being its own type, but with the gauge system change I'm not sure
this information is even needed anymore, but it can't hurt to keep it in
there.

I added a counter of cooldown drift and GCD clips at the end of the
timeline, mostly to sanity check. Still would really like to improve the
way oGCD weaves are handled (because burst is a mess atm) but for now I
think it's good enough.

Some GCD speeds also don't align properly / align weirdly to refresh
Thunder on time due to Thunderhead, this is not handled well for now.

I also just don't bother to check for GCD clips at level 70 because
there's so few weave slots to begin with.