Dialog for modifying sync info of an existing sheet (#736)
Closes #727 

Creates an equivalent of the save-as dialog but modifies in place. It
will reload the sheet since we need to invalidate all of the existing
GearItems.

This UI can be accessed via "More Actions", *or* by clicking on the
lvl/ilvl sync pill (it's a clickable button now).

The actual dialog uses the same UI as the "Save As" dialog.